Revert "Revert "Merge pull request #1773 from artsy/v2-refactor""
This reverts commit abaf8325a0580a3eec1a9567903659aa6dc5bfd0.
